# api/control/v1/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Caelis Control Client API",
    "version": "1.0.0-m3"
  },
  "servers": [{"url": "/api/control/v1"}],
  "paths": {
    "/initialize": {
      "get": {
        "operationId": "initializeClient",
        "responses": {
          "200": {"description": "Control protocol compatibility handshake", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ServerInfo"}}}},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "409": {"$ref": "#/components/responses/Conflict"},
          "500": {"$ref": "#/components/responses/InternalServerError"},
          "503": {"$ref": "#/components/responses/ServiceUnavailable"}
        }
      }
    },
    "/host/shutdown": {
      "post": {
        "operationId": "shutdownHost",
        "responses": {
          "200": {"description": "Host shutdown accepted", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HostStatus"}}}},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "409": {"$ref": "#/components/responses/Conflict"},
          "500": {"$ref": "#/components/responses/InternalServerError"},
          "503": {"$ref": "#/components/responses/ServiceUnavailable"}
        }
      }
    },
    "/sessions": {
      "get": {
        "operationId": "listSessions",
        "parameters": [
          {"name": "workspace_key", "in": "query", "schema": {"type": "string"}},
          {"name": "cwd", "in": "query", "schema": {"type": "string"}},
          {"name": "cursor", "in": "query", "schema": {"type": "string"}},
          {"name": "limit", "in": "query", "schema": {"type": "integer", "minimum": 1}}
        ],
        "responses": {
          "200": {"description": "Authorized Sessions", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SessionList"}}}},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "409": {"$ref": "#/components/responses/Conflict"},
          "500": {"$ref": "#/components/responses/InternalServerError"},
          "503": {"$ref": "#/components/responses/ServiceUnavailable"}
        }
      },
      "post": {
        "operationId": "createSession",
        "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateSessionRequest"}}}},
        "responses": {
          "200": {"$ref": "#/components/responses/CommandCommitted"},
          "202": {"$ref": "#/components/responses/CommandAccepted"},
          "400": {"$ref": "#/components/responses/CommandBadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "409": {"$ref": "#/components/responses/CommandConflict"},
          "500": {"$ref": "#/components/responses/InternalServerError"},
          "503": {"$ref": "#/components/responses/ServiceUnavailable"}
        }
      }
    },
    "/sessions/{session_id}": {
      "delete": {
        "operationId": "closeSession",
        "parameters": [
          {"$ref": "#/components/parameters/SessionID"},
          {"$ref": "#/components/parameters/IdempotencyKey"},
          {"$ref": "#/components/parameters/IfMatch"}
        ],
        "requestBody": {"$ref": "#/components/requestBodies/CloseSession"},
        "responses": {
          "200": {"$ref": "#/components/responses/CommandCommitted"},
          "202": {"$ref": "#/components/responses/CommandAccepted"},
          "400": {"$ref": "#/components/responses/CommandBadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "409": {"$ref": "#/components/responses/CommandConflict"},
          "500": {"$ref": "#/components/responses/InternalServerError"},
          "503": {"$ref": "#/components/responses/ServiceUnavailable"}
        }
      }
    },
    "/sessions/{session_id}/compact": {
      "post": {
        "operationId": "compactSession",
        "parameters": [
          {"$ref": "#/components/parameters/SessionID"},
          {"$ref": "#/components/parameters/IdempotencyKey"},
          {"$ref": "#/components/parameters/IfMatch"}
        ],
        "requestBody": {"$ref": "#/components/requestBodies/CompactSession"},
        "responses": {
          "200": {"$ref": "#/components/responses/CommandCommitted"},
          "202": {"$ref": "#/components/responses/CommandAccepted"},
          "400": {"$ref": "#/components/responses/CommandBadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "409": {"$ref": "#/components/responses/CommandConflict"},
          "500": {"$ref": "#/components/responses/InternalServerError"},
          "503": {"$ref": "#/components/responses/ServiceUnavailable"}
        }
      }
    },
    "/sessions/{session_id}/state": {
      "get": {
        "operationId": "getSessionState",
        "parameters": [{"$ref": "#/components/parameters/SessionID"}],
        "responses": {
          "200": {"description": "Reconnect bootstrap", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SessionState"}}}},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "409": {"$ref": "#/components/responses/Conflict"},
          "500": {"$ref": "#/components/responses/InternalServerError"},
          "503": {"$ref": "#/components/responses/ServiceUnavailable"}
        }
      }
    },
    "/sessions/{session_id}/status": {
      "get": {
        "operationId": "getSessionStatus",
        "parameters": [
          {"$ref": "#/components/parameters/SessionID"},
          {"name": "workspace_key", "in": "query", "schema": {"type": "string"}},
          {"name": "cwd", "in": "query", "schema": {"type": "string"}},
          {"name": "surface", "in": "query", "schema": {"type": "string"}},
          {"name": "diagnostics", "in": "query", "schema": {"type": "boolean"}},
          {"name": "workspace_trust_requirement", "in": "query", "schema": {"type": "boolean"}}
        ],
        "responses": {
          "200": {"description": "Session-scoped product status", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/StatusSnapshot"}}}},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "409": {"$ref": "#/components/responses/Conflict"},
          "500": {"$ref": "#/components/responses/InternalServerError"},
          "503": {"$ref": "#/components/responses/ServiceUnavailable"}
        }
      }
    },
    "/sessions/{session_id}/reconnect": {
      "get": {
        "operationId": "reconnectSession",
        "parameters": [
          {"$ref": "#/components/parameters/SessionID"},
          {"$ref": "#/components/parameters/After"},
          {"name": "Last-Event-ID", "in": "header", "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {
            "description": "Atomic SessionState bootstrap followed by typed append or transactional replacement deliveries.",
            "content": {"text/event-stream": {"schema": {"type": "string"}}}
          },
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "409": {"$ref": "#/components/responses/Conflict"},
          "500": {"$ref": "#/components/responses/InternalServerError"},
          "503": {"$ref": "#/components/responses/ServiceUnavailable"}
        }
      }
    },
    "/sessions/{session_id}/prompt": {
      "post": {
        "operationId": "promptSession",
        "parameters": [{"$ref": "#/components/parameters/SessionID"}, {"$ref": "#/components/parameters/IdempotencyKey"}, {"$ref": "#/components/parameters/IfMatch"}],
        "requestBody": {"$ref": "#/components/requestBodies/Prompt"},
        "responses": {
          "200": {"$ref": "#/components/responses/CommandCommitted"},
          "202": {"$ref": "#/components/responses/CommandAccepted"},
          "400": {"$ref": "#/components/responses/CommandBadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "409": {"$ref": "#/components/responses/CommandConflict"},
          "500": {"$ref": "#/components/responses/InternalServerError"},
          "503": {"$ref": "#/components/responses/ServiceUnavailable"}
        }
      }
    },
    "/sessions/{session_id}/steer": {
      "post": {
        "operationId": "steerSession",
        "parameters": [{"$ref": "#/components/parameters/SessionID"}, {"$ref": "#/components/parameters/IdempotencyKey"}, {"$ref": "#/components/parameters/IfMatch"}],
        "requestBody": {"$ref": "#/components/requestBodies/Steer"},
        "responses": {
          "200": {"$ref": "#/components/responses/CommandCommitted"},
          "202": {"$ref": "#/components/responses/CommandAccepted"},
          "400": {"$ref": "#/components/responses/CommandBadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "409": {"$ref": "#/components/responses/CommandConflict"},
          "500": {"$ref": "#/components/responses/InternalServerError"},
          "503": {"$ref": "#/components/responses/ServiceUnavailable"}
        }
      }
    },
...
    "/sessions/{session_id}/tasks": {
      "get": {
        "operationId": "listSessionTasks",
        "parameters": [{"$ref": "#/components/parameters/SessionID"}],
        "responses": {
          "200": {"description": "Authorized Task directory for one Session", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TaskList"}}}},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "409": {"$ref": "#/components/responses/Conflict"},
          "500": {"$ref": "#/components/responses/InternalServerError"},
          "503": {"$ref": "#/components/responses/ServiceUnavailable"}
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {"bearerAuth": {"type": "http", "scheme": "bearer"}},
    "parameters": {
...
      "IfMatch": {"name": "If-Match", "in": "header", "schema": {"type": "string"}},
...
    },
...
    "responses": {
...
      "BadRequest": {"description": "Malformed request", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}},
      "Unauthorized": {
        "description": "Authentication required or failed",
        "headers": {"WWW-Authenticate": {"schema": {"type": "string"}}},
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}
      },
      "Forbidden": {"description": "Authenticated principal is not authorized", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}},
...
      "Conflict": {"description": "Revision, idempotency, or state conflict", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}},
      "InternalServerError": {"description": "Internal server error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}},
...
    },
    "schemas": {
...
      "ServerInfo": {"type": "object", "additionalProperties": false, "required": ["protocol_version", "envelope_version", "api_version"], "properties": {"protocol_version": {"type": "integer"}, "envelope_version": {"type": "string"}, "api_version": {"type": "string"}, "distribution_version": {"type": "string"}, "build_id": {"type": "string"}, "build_kind": {"type": "string", "enum": ["dev", "release"]}, "server_id": {"type": "string"}, "instance_id": {"type": "string"}, "capabilities": {"type": "array", "items": {"type": "string"}}, "transports": {"type": "array", "items": {"type": "string"}}}},
...
    }
  },
  "security": [{"bearerAuth": []}]
}